+++ CHANGELOG.rst
@@ -8,7 +8,9 @@ CHANGELOG
 
 * bug-fix: Change module names to string type in __all__
 * feature: Save training output files in local mode
+* bug-fix: tensorflow-serving-api: SageMaker does not conflict with tensorflow-serving-api module version
 * feature: Local Mode: add support for local training data using file://
+* feature: Updated TensorFlow Serving api protobuf files
 
 1.2.4
 =====